ref(tech): rename main Checklist view to follow convention
Ref: #17
